Update vdiclient.py
I'm thinking of using this a network boot replacement for windows 10 -- because you can make windows 11 happy as vm on proxmox. I wanted the log in windows to appear in the middle of the screen. So I've added an import from yet another library which gives me screen dimensions, and added a location parameter to the login and vm's window. It's not sophisticated. but it seemed to work for me.